openwebuiapi.py created
