+++ server/run_openwebui.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+uvicorn --app-dir=src colette.openwebuiapi:app --port 8889 $*
